[Switchboard] Call deploymentConsumer after deployment has been stopped

+++ switchboard/src/main/java/nl/knaw/meertens/clariah/vre/switchboard/poll/PollServiceImpl.java
@@ -147,7 +147,7 @@ public class PollServiceImpl implements PollService {
   }
 
   private int calculateNewInterval(int previousInterval) {
-    return (int) Math.ceil(previousInterval * INCREASE_INTERVAL_FACTOR);
+    return (int) Math.ceil(((long) previousInterval) * INCREASE_INTERVAL_FACTOR);
   }
 
   private DeploymentStatusResponseDto requestDeploymentStatusReport(URI uri) {
